+++ frontend/common/src/components/form/MultiSelectField/MultiSelectField.test.tsx
@@ -13,7 +13,7 @@ import {
 import { FormProvider, useForm } from "react-hook-form";
 import type { FieldValues, SubmitHandler } from "react-hook-form";
 import IntlProvider from "react-intl/src/components/provider";
-import MultiSelectFieldV2 from "./MultiSelectFieldV2";
+import MultiSelectField from "./MultiSelectField";
 
 const Providers = ({
   children,
@@ -67,9 +67,9 @@ const renderWithProviders = (
     ...options,
   });
 
-describe("MultiSelectFieldV2", () => {
+describe("MultiSelectField", () => {
   it("should render properly with only label prop", () => {
-    renderWithProviders(<MultiSelectFieldV2 label="Foo Bar" />);
+    renderWithProviders(<MultiSelectField label="Foo Bar" />);
     expect(
       screen.getByRole("combobox", { name: /foo bar/i }),
     ).toBeInTheDocument();
@@ -85,7 +85,7 @@ describe("MultiSelectFieldV2", () => {
 
   it("should submit undefine when no default (no validation rules)", async () => {
     const mockSubmit = jest.fn();
-    renderWithProviders(<MultiSelectFieldV2 label="Foo Bar" options={[]} />, {
+    renderWithProviders(<MultiSelectField label="Foo Bar" options={[]} />, {
       wrapperProps: {
         onSubmit: mockSubmit,
       },
@@ -101,7 +101,7 @@ describe("MultiSelectFieldV2", () => {
   it("should submit correctly as array of values", async () => {
     const mockSubmit = jest.fn();
     renderWithProviders(
-      <MultiSelectFieldV2
+      <MultiSelectField
         label="Foo Bar"
         options={[
           { value: "BAZ", label: "Baz" },
@@ -147,7 +147,7 @@ describe("MultiSelectFieldV2", () => {
 
   it("should be clearable even when required", () => {
     renderWithProviders(
-      <MultiSelectFieldV2
+      <MultiSelectField
         label="Foo Bar"
         rules={{ required: "Required!" }}
         options={[{ value: "BAZ", label: "Baz" }]}
@@ -162,7 +162,7 @@ describe("MultiSelectFieldV2", () => {
   });
 
   it("should show loading indicator when isLoading", () => {
-    renderWithProviders(<MultiSelectFieldV2 label="Foo Bar" isLoading />);
+    renderWithProviders(<MultiSelectField label="Foo Bar" isLoading />);
 
     const loadingIndicator = document.querySelector(
       `.${CLASS_PREFIX}__loading-indicator`,
